A bunch of small code cleanups, mostly removal of old-skool casts, thanks to Dimitry Erastov

+++ com/pff/PSTRecipient.java
@@ -128,7 +128,7 @@ public class PSTRecipient {
 
 			if ( item.entryValueType == 0x0002 ) {
 				short s = (short)item.entryValueReference;
-				return (int)s;
+				return s;
 			}
 		}
